refactor: extract podcast search section component
